采购入库项目相关代码

+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/controller/admin/productbatch/vo/ErpProductBatchRespVO.java
@@ -37,7 +37,7 @@ public class ErpProductBatchRespVO {
      */
     @Schema(description = "批次入库数量")
     @ExcelProperty("批次入库数量")
-    private Integer inventoryQuantity;
+    private BigDecimal inventoryQuantity;
 
     @Schema(description = "批次编号")
     @ExcelProperty("批次编号")
