+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/dal/dataobject/productbatch/ErpProductBatchDO.java
@@ -53,7 +53,7 @@ public class ErpProductBatchDO extends BaseDO {
     /**
      * 批次入库数量
      */
-    private Integer inventoryQuantity;
+    private BigDecimal inventoryQuantity;
     /**
      * 备注
      */
